mejorando el modelo, incluyendo algunos patch endpoint, y validando operacion update/patch con la inclusión de algunos mensajes internacionalizados

+++ AppCuentas/src/main/java/com/jislas/devsu/appcuentas/models/dto/movimiento/MovimientoDto.java
@@ -3,6 +3,7 @@ package com.jislas.devsu.appcuentas.models.dto.movimiento;
 import com.jislas.devsu.appcuentas.models.entity.TipoMovimiento;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -10,6 +11,6 @@ public class MovimientoDto {
     private Long id;
     private LocalDate fecha;
     private TipoMovimiento tipoMovimiento;
-    private double valor;
-    private double saldo;
+    private BigDecimal valor;
+    private BigDecimal saldo;
 }
